+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v2/validator/GetNewBlockV2Test.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 ConsenSys AG.
+ * Copyright 2021 ConsenSys AG.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.beaconrestapi.handlers.v1.validator;
+package tech.pegasys.teku.beaconrestapi.handlers.v2.validator;
 
 import static javax.servlet.http.HttpServletResponse.SC_BAD_REQUEST;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -30,36 +30,34 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.ArgumentMatcher;
-import org.mockito.ArgumentMatchers;
+import tech.pegasys.teku.api.SchemaObjectProvider;
 import tech.pegasys.teku.api.ValidatorDataProvider;
-import tech.pegasys.teku.api.response.v1.validator.GetNewBlockResponse;
+import tech.pegasys.teku.api.response.v2.validator.GetNewBlockResponseV2;
 import tech.pegasys.teku.api.schema.BLSSignature;
 import tech.pegasys.teku.api.schema.BeaconBlock;
-import tech.pegasys.teku.beaconrestapi.RestApiConstants;
-import tech.pegasys.teku.beaconrestapi.schema.BadRequest;
 import tech.pegasys.teku.bls.BLSTestUtil;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.provider.JsonProvider;
+import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecMilestone;
+import tech.pegasys.teku.spec.TestSpecFactory;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
-public class GetNewBlockTest {
+public class GetNewBlockV2Test {
   private final tech.pegasys.teku.bls.BLSSignature signatureInternal =
       BLSTestUtil.randomSignature(1234);
   private BLSSignature signature = new BLSSignature(signatureInternal);
   private Context context = mock(Context.class);
-  protected final ValidatorDataProvider provider = mock(ValidatorDataProvider.class);
-  protected final JsonProvider jsonProvider = new JsonProvider();
+  private final ValidatorDataProvider provider = mock(ValidatorDataProvider.class);
+  private final JsonProvider jsonProvider = new JsonProvider();
   private GetNewBlock handler;
-  private final DataStructureUtil dataStructureUtil = new DataStructureUtil();
-  private final Bytes32 graffiti = dataStructureUtil.randomBytes32();
+  private final Spec spec = TestSpecFactory.createMinimalAltair();
+  private final DataStructureUtil dataStructureUtil = new DataStructureUtil(spec);
 
   @SuppressWarnings("unchecked")
   final ArgumentCaptor<SafeFuture<String>> args = ArgumentCaptor.forClass(SafeFuture.class);
@@ -70,56 +68,29 @@ public class GetNewBlockTest {
   }
 
   @Test
-  void shouldRequireThatRandaoRevealIsSet() throws Exception {
-    badRequestParamsTest(Map.of(), "'randao_reveal' cannot be null or empty.");
-  }
-
-  @Test
   void shouldReturnBlockWithoutGraffiti() throws Exception {
     final Map<String, String> pathParams = Map.of(SLOT, "1");
+    final SchemaObjectProvider schemaProvider = new SchemaObjectProvider(spec);
     final Map<String, List<String>> queryParams =
         Map.of(RANDAO_REVEAL, List.of(signature.toHexString()));
-    Optional<BeaconBlock> optionalBeaconBlock =
-        Optional.of(
-            new BeaconBlock(dataStructureUtil.randomBeaconBlock(dataStructureUtil.randomLong())));
+
+    final tech.pegasys.teku.spec.datastructures.blocks.BeaconBlock randomBeaconBlock =
+        dataStructureUtil.randomBeaconBlock(dataStructureUtil.randomLong());
+
+    final BeaconBlock altairBlock = schemaProvider.getBeaconBlock(randomBeaconBlock);
     when(context.queryParamMap()).thenReturn(queryParams);
     when(context.pathParamMap()).thenReturn(pathParams);
-    when(provider.getMilestoneAtSlot(UInt64.ONE)).thenReturn(SpecMilestone.PHASE0);
+    when(provider.getMilestoneAtSlot(UInt64.ONE)).thenReturn(SpecMilestone.ALTAIR);
     when(provider.getUnsignedBeaconBlockAtSlot(ONE, signature, Optional.empty()))
-        .thenReturn(SafeFuture.completedFuture(optionalBeaconBlock));
+        .thenReturn(SafeFuture.completedFuture(Optional.of(altairBlock)));
     handler.handle(context);
 
     verify(context).result(args.capture());
     SafeFuture<String> result = args.getValue();
     assertThat(result)
         .isCompletedWithValue(
-            jsonProvider.objectToJSON(new GetNewBlockResponse(optionalBeaconBlock.get())));
-  }
-
-  @Test
-  void shouldReturnBlockWithGraffiti() throws Exception {
-    final Map<String, List<String>> params =
-        Map.of(
-            RANDAO_REVEAL,
-            List.of(signature.toHexString()),
-            RestApiConstants.GRAFFITI,
-            List.of(graffiti.toHexString()));
-    Optional<BeaconBlock> optionalBeaconBlock =
-        Optional.of(
-            new BeaconBlock(dataStructureUtil.randomBeaconBlock(dataStructureUtil.randomLong())));
-    when(context.queryParamMap()).thenReturn(params);
-    when(context.pathParamMap()).thenReturn(Map.of(SLOT, "1"));
-    when(provider.getMilestoneAtSlot(UInt64.ONE)).thenReturn(SpecMilestone.PHASE0);
-    when(provider.getUnsignedBeaconBlockAtSlot(ONE, signature, Optional.of(graffiti)))
-        .thenReturn(SafeFuture.completedFuture(optionalBeaconBlock));
-    handler.handle(context);
-
-    verify(context).result(args.capture());
-    SafeFuture<String> result = args.getValue();
-
-    assertThat(result)
-        .isCompletedWithValue(
-            jsonProvider.objectToJSON(new GetNewBlockResponse(optionalBeaconBlock.get())));
+            jsonProvider.objectToJSON(
+                new GetNewBlockResponseV2(SpecMilestone.ALTAIR, altairBlock)));
   }
 
   @Test
@@ -153,32 +124,5 @@ public class GetNewBlockTest {
     handler.handle(context);
 
     verify(context).status(SC_BAD_REQUEST);
-  }
-
-  @Test
-  void shouldReturnBadRequestForAltairBlocks() throws Exception {
-    final Map<String, List<String>> params =
-        Map.of(RANDAO_REVEAL, List.of(signature.toHexString()));
-    when(context.pathParamMap()).thenReturn(Map.of(SLOT, "1"));
-    when(context.queryParamMap()).thenReturn(params);
-    when(provider.getMilestoneAtSlot(UInt64.ONE)).thenReturn(SpecMilestone.ALTAIR);
-    handler.handle(context);
-
-    verify(context).status(SC_BAD_REQUEST);
-    verify(context).result(ArgumentMatchers.contains("please fetch via /eth/v2/validator/blocks"));
-  }
-
-  private void badRequestParamsTest(final Map<String, List<String>> params, String message)
-      throws Exception {
-    when(context.queryParamMap()).thenReturn(params);
-    when(context.pathParamMap()).thenReturn(Map.of(SLOT, "1"));
-
-    handler.handle(context);
-    verify(context).status(SC_BAD_REQUEST);
-
-    if (StringUtils.isNotEmpty(message)) {
-      BadRequest badRequest = new BadRequest(message);
-      verify(context).result(jsonProvider.objectToJSON(badRequest));
-    }
   }
 }